% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LakeMapObjects.R
\name{basemap_lake259}
\alias{basemap_lake259}
\title{basemap_lake259}
\format{
a ggplot2 baseplot of Lake 259
}
\description{
Lake 259 shoreline as ggplot base plot (geom_sf)
}
\examples{
\dontrun{
data("base_lake259")
basemap_lake259
}
}
\seealso{
shape_lake259
}
