% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LakeMapObjects.R
\name{basemap_lake379}
\alias{basemap_lake379}
\title{basemap_lake379}
\format{
a ggplot2 baseplot of Lake 379
}
\description{
Lake 379 shoreline as ggplot base plot (geom_sf)
}
\examples{
\dontrun{
data("base_lake379")
basemap_lake379
}
}
\seealso{
shape_lake379
}
