% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LakeMapObjects.R
\name{basemap_lake622}
\alias{basemap_lake622}
\title{basemap_lake622}
\format{
a ggplot2 baseplot of Lake 622
}
\description{
Lake 622 shoreline as ggplot base plot (geom_sf)
}
\examples{
\dontrun{
data("base_lake622")
basemap_lake622
}
}
\seealso{
shape_lake622
}
