% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LakeMapObjects.R
\name{basemap_lake623}
\alias{basemap_lake623}
\title{basemap_lake623}
\format{
a ggplot2 baseplot of Lake 623
}
\description{
Lake 623 shoreline as ggplot base plot (geom_sf)
}
\examples{
\dontrun{
data("base_lake623")
basemap_lake623
}
}
\seealso{
shape_lake623
}
