% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ELA_data.R
\name{export_ELA_data}
\alias{export_ELA_data}
\title{export_ELA_data}
\usage{
export_ELA_data(evfile, calfile)
}
\arguments{
\item{evfile}{The name and file path to the required ev file}

\item{calfile}{The name and file path to the required calibration file}
}
\value{
Returns usethis messages to provide status of progress of COM application
}
\description{
This function provides COM control over Echoview Software. A user must have all of the necessary licenses and permissions to run Echoview.
}
\examples{
\dontrun{
export_ELA_data("myev.EV")
all_ev_files <- dir(recursive = TRUE, pattern = "EV$", full.names = TRUE)
lapply(all_ev_files, export_ELA_data)
}
}
