% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LakeMapObjects.R
\name{shape_lake259}
\alias{shape_lake259}
\title{shape_lake259}
\format{
an sf spatial data frame
}
\description{
Lake 259 shoreline as shapefile
}
\examples{
\dontrun{
data("shape_lake259")
class(shape_lake259)
plot(shape_lake259)
}
}
\seealso{
basemap_lake259
}
