% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LakeMapObjects.R
\name{shape_lake379}
\alias{shape_lake379}
\title{shape_lake379}
\format{
an sf spatial data frame
}
\description{
Lake 379 shoreline as shapefile
}
\examples{
\dontrun{
data("shape_lake379")
class(shape_lake379)
plot(shape_lake379)
}
}
\seealso{
basemap_lake379
}
