% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LakeMapObjects.R
\name{shape_lake622}
\alias{shape_lake622}
\title{shape_lake622}
\format{
an sf spatial data frame
}
\description{
Lake 622 shoreline as shapefile
}
\examples{
\dontrun{
data("shape_lake622")
class(shape_lake622)
plot(shape_lake622)
}
}
\seealso{
basemap_lake622
}
