% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LakeMapObjects.R
\name{shape_lake623}
\alias{shape_lake623}
\title{shape_lake623}
\format{
an sf spatial data frame
}
\description{
Lake 623 shoreline as shapefile
}
\examples{
\dontrun{
data("shape_lake623")
class(shape_lake623)
plot(shape_lake623)
}
}
\seealso{
basemap_lake623
}
